implemented GET and POST services for classes: Survey, Question, Answer

+++ academy-survey/src/main/java/lt/daivospakalikai/academysurvey/Survey/SurveyRepository.java
@@ -2,6 +2,6 @@ package lt.daivospakalikai.academysurvey.Survey;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SurveyRepository extends JpaRepository<Survey, Long> {
+public interface SurveyRepository extends JpaRepository<Survey, Integer> {
 
 }
